requirements.txt
torch
torchvision
opencv-python
fastapi
uvicorn
pillow
numpy
